refactor(FE): queries를 각 routes 안으로 이동, 외부로 export 하는 디렉토리 추가
