+++ frontend/src/routes/notices/$id/index.tsx
@@ -2,7 +2,7 @@ import { useSuspenseQuery } from '@tanstack/react-query'
 import { createFileRoute } from '@tanstack/react-router'
 import { Flex } from '~/components/layout/flex'
 import { Heading } from '~/components/typo/heading'
-import { getNoticeDetailQueryOptions } from '~/queries/notices/notices.query'
+import { getNoticeDetailQueryOptions } from '~/routes/notices/-queries/notices.query'
 import { NOTICE_CATEGORY_OPTIONS } from '../-constants/notice-category-options'
 import { Text } from '~/components/typo/text'
 import { Tag } from '~/components/ui/tag'
