+++ frontend/src/routes/notices/-components/list/notice-results-section.tsx
@@ -4,7 +4,7 @@ import { Route } from '~/routes/notices'
 import { Box } from '~/components/layout/box'
 import { Grid } from '~/components/layout/grid'
 import { useSuspenseQuery } from '@tanstack/react-query'
-import { getNoticesQueryOptions } from '~/queries/notices/notices.query'
+import { getNoticesQueryOptions } from '../../-queries/notices.query'
 import { NoticeCard } from './notice-card'
 import { Pagination } from '~/components/ui/pagination'
 import { useNavigate } from '@tanstack/react-router'
