+++ frontend/src/routes/notices/-exports/components.ts
@@ -1,0 +1,1 @@
+export { NoticeCard } from '../-components/list/notice-card'
