+++ frontend/src/routes/notices/-exports/queries.ts
@@ -1,0 +1,1 @@
+export { getNoticesQueryOptions } from '../-queries/notices.query'
