+++ frontend/src/routes/notices/-queries/notices.model.ts
