+++ frontend/src/routes/policies/-components/list/policy-card.tsx
@@ -4,7 +4,7 @@ import { Flex } from '~/components/layout/flex'
 import { Dot } from '~/components/ui/dot'
 import { css } from '@emotion/react'
 import { Button } from '~/components/ui/button'
-import type { PoliciesResponse } from '~/queries/policies/policies.type'
+import type { PoliciesResponse } from '~/routes/policies/-queries/policies.type'
 import { POLICY_CATEGORY_OPTIONS } from '../../-constants/policy-category-options'
 import type { Prettify } from '~/types/prettify'
 import { Link } from '@tanstack/react-router'
