+++ frontend/src/routes/policies/-queries/policies.query.ts
@@ -1,6 +1,6 @@
 import { queryOptions } from '@tanstack/react-query'
 import { POLICIES_ENDPOINT, POLICIES_KEY } from './policies.model'
-import { request } from '../request'
+import { request } from '../../../utils/request'
 import type { PoliciesResponse, Policy } from './policies.type'
 
 export const getPoliciesQueryOptions = () =>
